:pencil2: meeting-reminder: Reword for style
A time is usually followed by the word "Uhr" in written German.  Numbers
and units are not abbreviated in continous text.  Personal pronouns are
written with capital letter when directly addressing someone, this is
more polite.  And finally: when starting a sentence with plural, stay in
plural and do not switch to singular (this removes a typo as well).